refactor(observability-otlp): remove OpenTelemetry SDK dependencies
Move HyperOtelClient, ClientError, and build_tonic_channel from src/client.rs
into transport/http_client.rs and drop the opentelemetry_http HttpClient
impl along with the opentelemetry-http and async-trait dependencies. No
opentelemetry* SDK crate remains; the transport owns all client logic.
