+++ modules/observability-otlp/Cargo.toml
@@ -16,7 +16,6 @@ tonic = "0.14"
 rustls = { version = "0.23", default-features = false, features = ["std", "aws_lc_rs"] }
 rustls-native-certs = "0.8"
 webpki-roots = "1.0"
-async-trait = "0.1"
 bytes = "1"
 http = "1"
 http-body-util = "0.1"
@@ -32,12 +31,6 @@ base64 = "0.23"
 hex = "0.4"
 rand = "0.10"
 crossbeam-queue = "0.3.13"
-
-# TODO: remove opentelemetry-http below after the transport owns the HTTP
-# client logic (Step 6).
-opentelemetry-http = { version = "0.32.0", default-features = false, features = [
-  "hyper"
-]}
 
 [build-dependencies]
 find_git = "1.2.1"
