+++ modules/observability-otlp/src/transport/mod.rs
@@ -4,4 +4,5 @@
 pub mod client;
 pub mod grpc;
 pub mod http;
+pub mod http_client;
 pub mod json;
